1.0.0-RC1

+++ vertigo-dynamo-impl/src/main/java/io/vertigo/dynamo/impl/store/datastore/DataStoreConfigImpl.java
@@ -37,7 +37,7 @@ public final class DataStoreConfigImpl implements DataStoreConfig {
 	private final LogicalDataStoreConfig logicalDataStoreConfig;
 
 	/**
-	 * Constructeur.
+	 * Constructor.
 	 * @param dataStorePlugins DataStorePlugins list
 	 * @param cacheManager Manager de gestion du cache
 	 */
@@ -57,7 +57,7 @@ public final class DataStoreConfigImpl implements DataStoreConfig {
 	 * @param serializeElements Set if elements should be serialized or not (serialization guarantee elements are cloned and not modified)
 	 */
 	@Override
-	public void registerCacheable(final DtDefinition dtDefinition, final long timeToLiveInSeconds, final boolean isReloadedByList, final boolean serializeElements) {
+	public void registerCacheable(final DtDefinition dtDefinition, final int timeToLiveInSeconds, final boolean isReloadedByList, final boolean serializeElements) {
 		cacheStoreConfig.registerCacheable(dtDefinition, timeToLiveInSeconds, isReloadedByList, serializeElements);
 	}
